Move OC.Contactsmenu and OC.Backbone to the server bundle
