+++ core/js/core.json
@@ -9,8 +9,6 @@
 		"jquery.ocdialog.js",
 		"oc-dialogs.js",
 		"js.js",
-		"oc-backbone.js",
-		"oc-backbone-webdav.js",
 		"l10n.js",
 		"share.js",
 		"sharetemplates.js",
@@ -22,7 +20,6 @@
 		"sharedialogresharerinfoview.js",
 		"sharedialogshareelistview.js",
 		"octemplate.js",
-		"contactsmenu.js",
 		"contactsmenu_templates.js",
 		"eventsource.js",
 		"config.js",
