+++ core/js/merged-template-prepend.json
@@ -12,7 +12,6 @@
   "oc-requesttoken.js",
   "mimetype.js",
   "mimetypelist.js",
-  "oc-backbone.js",
   "select2-toggleselect.js",
   "placeholder.js",
   "jquery.avatar.js",
